[~] refactor code
