+++ effectiveOfficeBackend/src/main/kotlin/office/effective/common/notifications/NotificationSender.kt
@@ -1,0 +1,8 @@
+package office.effective.common.notifications
+
+class NotificationSender {
+
+    fun sendChangeNotification(processId: Int, channelName: String, payload: String) {
+        //todo: send FCM message
+    }
+}
